loading done
Need to actually get everything finished now and stop messing with components. Imported headless UI to assist with this. I still want to do styling myself. I think the loading bar was a huge waste of time but it was at least moderately entertaining. Spending two days on it was absolutely stupid though.